+++ module-apps/application-settings-new/ApplicationSettings.hpp
@@ -25,7 +25,7 @@ namespace gui::window::name
 
     inline const std::string display_and_keypad = "DisplayAndKeypad";
 
-    inline const std::string change_settings = "ChangeSettings";
+    inline const std::string switch_sim      = "SwitchSim";
     inline const std::string all_operators   = "AllOperators";
     inline const std::string import_contacts = "ImportContacts";