+++ changelog.md
@@ -13,6 +13,7 @@
 * `[audio]` Enabled sounds priorities, looping and merging
 * `[renderer]` Add drawing circles and arcs functionality.
 * `[bluetooth]` Added scan and pairing support.
+* `[settings]` Added sim setup window.
 
 ### Changed
 * `[special input]` Added emoji selector and changed special character selector.